requirements.txt
# Core dependencies
pyobjc-core>=9.0
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-Vision>=9.0

# Notion API
notion-client>=2.0.0

# OCR fallback
pytesseract>=0.3.10
Pillow>=10.0.0

# CLI
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2

